# Conflicts:
#	api/include/opentelemetry/trace/trace_state.h
#	api/test/trace/trace_state_test.cc

+++ sdk/test/trace/attribute_utils_test.cc
@@ -22,6 +22,6 @@ TEST(AttributeMapTest, AttributesConstruction)
 
   for (int i = 0; i < kNumAttributes; i++)
   {
-    EXPECT_EQ(opentelemetry::nostd::get<int64_t>(map.GetAttributes().at(keys[i])), values[i]);
+    EXPECT_EQ(opentelemetry::nostd::get<int>(map.GetAttributes().at(keys[i])), values[i]);
   }
 }
